Add files via upload
Description
Your program is able to play using one card. Let's make our game serious and implement a set of cards now!

Let the user decide how many cards they would like to keep. First, ask the player to enter the desired number of cards. Then, ask to input the term and the definition of every card. If the initial number was 100, then... well, what a pity!

In the end, when all of the cards are defined and saved, your program is finally ready to play! Question the player about all the new words they have entered. The program should give the term and ask for a definition. Let the game begin!

How to submit?
This stage is auto-graded. The grader will behave as in example, so you can change an internal logic of your code, but the output should be similar. Here are some important notes:

When you ask for a definition of a card, you should write the term of the card in quotes. Example: Print the definition of "black":.
If the answer is correct, you should print Correct answer.. And if the answer is wrong, you should print Wrong answer. The correct one is "black"., where "black" is the correct definition.
Example
The symbol > represents the user input. Notice that it's not the part of the input.

Input the number of cards:
> 2
The card #1:
> black
The definition of the card #1:
> white
The card #2:
> white
The definition of the card #2:
> black
Print the definition of "black":
> white
Correct answer.
Print the definition of "white":
> blue
Wrong answer. The correct one is "black".